Move from index all tests to test.js file - Append App child in content index.html coming form public/index.html directory
